Defer handling of formats before autoloading
If the SpeedDatingFormat command is used to alter list of supported
formats when autoloading hasn't yet occurred, just save the arguments
into a list for later processing.  This allows deferring the actual
loading until it SpeedDating is actually used, even if custom date
formats are being defined.

If that command is used to get information, either the current list of
formats or usage information, do autoloading and handle that request
immediately.