Fix localtime() in neovim
In neovim `reltime()` is based on the system uptime rather than the
epoch so it can't be used to get the local time. Use the same fallback
as is used for non-Unix systems instead. This means that %k won't
support milliseconds as added in d03fc1a, but at least larger units of
time should be correct.